Add definition of a Growth Plan
Addresses issue #50

A separate PR will address the fact that some of the currently listed
graduation criteria are subjective.  (For some history, when the policy
document was accepted by the TAC as a starting point back in October 2019,
there were unresolved questions/issues on the Graduation stage that were
punted for later discussion since no projects are proposed for that
stage yet.)  That's closely related to this one, but the intent is to
decouple the two as separate PRs.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>