Narrow list of example technologies
Currently the TAC does not take a position on whether a specific technology
is or is not confidential computing.  Similarly it does not claim to
have a complete list.

We are already on thin ice by calling out some technologies as
"claiming support" since some may still interpret this as endorsement of
said claims, so it would seem best to be very conservative about including
things if we continue to include them, and only refer to them as
examples.

There is apparently some debate around the strength of some technologies
(e.g., SEV as opposed to SEV-SNP), so this proposal is to only mention things
that are not debated so as to not take any position in the debate, while
adding the word "Examples" in the diagrams to make it clear it is not
intended to be a complete list.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>