Proposal to include attestation in CC definition
But not in TEE definition.  Our scoping.md document already says:
"The CCC position is that only hardware-rooted technologies can provide
the desired level of security, in particular for attestation."
But attestation is not in the current definition, it's listed as
optional, which has generated a lot of discussion since from people
who believe it needs to be there.

This PR is an attempt to see whether we can get consensus to add it
without changing the definition of a TEE.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>